Refactore GenerateReport and ReportFolderExist methods. Split the code from GenerateReport within new methods: IsReportGenerated() and WriteReportFile()
