+++ Export Analysis Reports/Sdl.Community.ExportAnalysisReports/Interfaces/IReportService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.ComponentModel;
 using System.Xml;
 using Sdl.Community.ExportAnalysisReports.Model;
 
@@ -12,7 +13,11 @@ namespace Sdl.Community.ExportAnalysisReports.Interfaces
 		void SaveExportPath(string reportOutputPath);
 		string GetJsonReportPath(string jsonPath);
 		bool IsSameReportPath(string reportOutputPath);
+		bool ReportFolderExist(XmlNode projectInfoNode, string projectXmlPath);
+
+		bool IsReportGenerated(string reportOutputPath, bool isChecked, OptionalInformation optionalInformation, BindingList<ProjectDetails> projects);
 
 		string JsonPath { get; set; }
+		string ReportsFolderPath { get; set; }
 	}
 }
